moved all models into shared/models folder, updated all imports
