+++ client/src/app/account-card/services/account.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable, tap } from 'rxjs';
-import { AccountModel } from '../account.model';
+import { AccountModel } from '../../shared/models/account.model';
 
 const URL = 'http://localhost:5000/api/myaccount';
