+++ client/src/app/shared/models/statistics.model.ts
