+++ client/src/app/transaction-card/services/transaction.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable, tap } from 'rxjs';
-import { TransactionModel } from '../transaction.model';
+import { TransactionModel } from '../../shared/models/transaction.model';
 
 const URL = 'http://localhost:5000/api/transactions';
